+++ src/Client/ClientCore/Systems/ClientNetwork/ClientNetworkSystem.cs
@@ -65,8 +65,7 @@ public sealed class ClientNetworkSystem : ISystem, IDisposable
         EventId.Client_Network_BeginConnection,
         EventId.Client_Network_RegisterAccount,
         EventId.Core_WorldManagement_Subscribe,
-        EventId.Core_WorldManagement_Unsubscribe,
-        EventId.Core_Movement_RequestMove
+        EventId.Core_WorldManagement_Unsubscribe
     };
 
     public int WorkloadEstimate => 20;